[JENKINS-46386] - Cleanup Tests
